backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
requests
pydantic
langchain
langchain-openai
